Fix RtD links and Update documentation links and naming.
Updated documentation links and clarified naming conventions in index.md and README.md. Changed the reference to the license from the generic term [license] to [Apache 2.0 license]. Additionally, the [command-line reference] has been renamed to [Command-line Usage], and [API Reference] was added for clarity. These changes improve the user's understanding of the content and ensure that the information is accurately linked.