Extend Dimension data model in mdio schema
The Dimension class in the mdio schema has been updated to not only include 'name' and 'size' fields with their respective descriptions, but also to introduce a new optional field 'chunksize'. This update enhances the data model, providing a clearer definition of a Dimension object and allowing for chunking specification.