Update noxfile.py dependencies
The nox.needs_version has been updated to a newer version and several changes have been made to the package install list in the noxfile.py script. sphinx-design and autodoc-pydantic have been added, and the git version of myst-nb has been replaced with a stable version. This enhances documentation generation with more stability and functionality.