Improved Dataset modeling in MDIO schema v1
The Dataset modeling in MDIO schema is refactored and improved. A new DatasetInfo class is created to hold the dataset's fundamental information. The current Dataset class's fields are updated, including replacing UserAttributes with the new DatasetInfo class for dataset metadata field. This makes the management of dataset information more efficient and logical.