New CLI Options (#364)
* Added utility copy and info commands.

* Improve formating

* Updates to pass pre-commit.

* Add access pattern to info.

* Add tests.

* Move utility commands to root level.

* Update linting and tests.

* Resolve PR comments.

* Fix mdio copy.

* Remove duplicate tmp in .gitignore.

* Remove unnecessary try/except block in segy.py

* Make 'info' command work with new CLI and add rich printing

* make copy work with new CLI

* Update description in info.py module

* Change input mdio option to argument in info command

* Update variable name and table title in info.py

* Replace copy command filename options with arguments

* make tests work for option -> argument conversion

* Refactor imports and command options in copy.py

* revert back to click types, better error handling

* Refactor segy.py and update test_main.py

Updated the segy.py file to import specific functions from click, rather than the entire module. The command decorator's function signatures and calls are also updated. This is to improve specificity and reduce unnecessary overhead. Additionally, modified the way command line arguments are passed in test_main.py as per the refactored changes in the main function.

* Add future annotations to copy command

* Refactor import location in copy.py

The import statement for 'MDIOReader' in the copy.py file has been moved to a more appropriate position. This change aims to maximize importing efficiency by having the import statement closer to where the imported module is being used.

* Refactor MDIO info command for better code organization

The MDIO info command is refactored to enhance code readability and maintenance. The new structure involves separate functions for 'cast_stats', 'parse_grid' and 'pretty_print' to each perform distinct tasks. This improves the clear segregation of tasks and ease of future modifications.

* Move pytest-dependency to test suite installs

* Add future annotations import to info.py

* Update usage documentation for mdio commands

The documentation for the mdio commands has been updated to reflect changes in the command syntax. Parameters for input and output files are now required positional arguments, rather than options, enhancing the clarity and readability of the commands.

* directly import click_params objects

* directly import click_params objects

* Add "fastentrypoints" to build requirements

* change overwrite to flag

---------

Co-authored-by: Mark Roberts <mark.roberts@tgs.com>