Fix broken CI: Make CI tooling constraint paths absolute (#382)
* Update pip and pipx syntax in test workflow

Updated the syntax used for specifying constraints during pip, poetry, and nox installations in the GitHub tests workflow. Shortened "--constraint" to "-c" for pip operations, and wrapped pip arguments in quotes for pipx operations for better readability and consistency.

* add github workspace prefix to file paths

* Refactor $GITHUB_WORKSPACE usage in test workflows

Updated all instances of $GITHUB_WORKSPACE usage in the .github/workflows/tests.yml file to ${GITHUB_WORKSPACE}. This ensures proper variable usage within the GitHub Actions environment. Fixed all occurrences in pip, pipx, Poetry, and Nox installation steps.

* Update GitHub workflow to use env variable

The commit replaces the use of `GITHUB_WORKSPACE` to `env.GITHUB_WORKSPACE` in various pip, pipx, and poetry install commands throughout the GitHub actions workflow. This is to ensure that the environmental variable is consistently used for accessing the workspace path.

* Update GitHub actions paths

The commit replaces the use of environment variables with the built-in GitHub context for specified paths. Changes were made in the 'tests.yml' workflow file, particularly on steps that upgrade pip, install Poetry and Nox. This is to ensure the uniformity and efficiency of the action scripts.